Refactor main loop
