+++ Makefile
@@ -82,6 +82,7 @@ SRC				= $(SRC_DIR)/main.c \
 					$(SRC_DIR)/parser/parser_utils.c \
 					$(SRC_DIR)/parser/parser.c \
 					$(SRC_DIR)/parser/try_parse_redirect.c \
+					$(SRC_DIR)/routine/loop.c \
 					$(SRC_DIR)/token/constructor.c \
 					$(SRC_DIR)/token/destructor.c \
 					$(SRC_DIR)/token/getter.c  \
