+++ include/routine.h
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   routine.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/10/19 16:13:20 by reasuke           #+#    #+#             */
+/*   Updated: 2024/10/19 16:13:33 by reasuke          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef ROUTINE_H
+# define ROUTINE_H
+
+# include "ctx.h"
+
+void	loop(t_ctx *ctx);
+
+#endif
